deploy 20-02-21

+++ src/main/java/com/main/glory/servicesImpl/AdminServciceImpl.java
@@ -6,7 +6,9 @@ import com.main.glory.Dao.admin.DepartmentDao;
 import com.main.glory.model.admin.ApprovedBy;
 import com.main.glory.model.admin.Company;
 import com.main.glory.model.admin.Department;
+import com.main.glory.model.dyeingSlip.DyeingSlipMast;
 import com.main.glory.model.jet.request.AddJet;
+import com.main.glory.model.user.UserData;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +18,9 @@ import java.util.List;
 public class AdminServciceImpl {
 
     @Autowired
+    DyeingSlipServiceImpl dyeingSlipService;
+
+    @Autowired
     DepartmentDao departmentDao;
 
     @Autowired
@@ -23,6 +28,9 @@ public class AdminServciceImpl {
 
     @Autowired
     ApproveByDao approveByDao;
+
+    @Autowired
+    UserServiceImpl userService;
 
 
     public void saveCompanyName(Company company) throws Exception {
@@ -70,21 +78,22 @@ public class AdminServciceImpl {
         }
     }
 
-    public Boolean deleteApprovedById(Long id) {
-        try {
+    public Boolean deleteApprovedById(Long id) throws Exception {
+
             ApprovedBy approvedByExist = approveByDao.getApprovedById(id);
             if (approvedByExist == null)
                 throw new Exception("no data found");
+
+            List<DyeingSlipMast> dyeingSlipMasts =dyeingSlipService.getDyeingSlipByApprovedId(id);
+            if(!dyeingSlipMasts.isEmpty())
+                throw new Exception("can't delete this record");
 
             approveByDao.deleteApprovedById(id);
             return true;
 
 
-        }
-        catch (Exception e)
-        {
-            return false;
-        }
+
+
     }
 
     public void saveDepartment(Department c) throws Exception {
@@ -104,6 +113,9 @@ public class AdminServciceImpl {
             Department exist = departmentDao.getDepartmentById(id);
             if (exist == null)
                 throw new Exception("no data found");
+            List<UserData> userDataList = userService.getAllUserByDepartment(exist.getName());
+            if(!userDataList.isEmpty())
+                throw new Exception("can't delete the department");
 
             departmentDao.deleteDepartmentById(id);
             return true;
@@ -118,5 +130,74 @@ public class AdminServciceImpl {
 
     public List<Department> getAllDepartmentList() {
         return departmentDao.getAllDepartment();
+    }
+
+    public Company getCompanyById(Long id) throws Exception {
+
+        Company c= companyDao.getCompanyById(id);
+        if(c==null)
+            throw new Exception("no company found");
+
+        return c;
+    }
+
+    public void updateCompany(Company company) throws Exception {
+        Company companyExist = companyDao.getCompanyById(company.getId());
+        if(companyExist==null) {
+            throw new Exception("no data found");
+        }
+
+        List<UserData> userDataList =userService.getAllUserByCompany(companyExist.getName());
+        companyDao.save(company);
+
+        for(UserData userData:userDataList)
+        {
+            userService.updateUserCompanyById(userData.getId(),company.getName());
+        }
+
+    }
+
+    public void updateApprovedBy(ApprovedBy approvedBy) throws Exception {
+        ApprovedBy approvedExist = approveByDao.getApprovedById(approvedBy.getId());
+        if (approvedExist==null)
+            throw new Exception("no record found");
+
+        List<DyeingSlipMast> dyeingSlipMasts = dyeingSlipService.getDyeingSlipByApprovedId(approvedBy.getId());
+        approveByDao.save(approvedBy);
+
+        for(DyeingSlipMast dyeingSlipMast:dyeingSlipMasts)
+        {
+            dyeingSlipService.updateDyeingSlipWithApproveById(approvedBy.getId(),dyeingSlipMast.getId());
+        }
+
+    }
+
+    public ApprovedBy getApprovedById(Long id) throws Exception {
+        ApprovedBy approvedByExist = approveByDao.getApprovedById(id);
+        if(approvedByExist==null)
+            throw new Exception("no data found");
+        return approvedByExist;
+    }
+
+    public Department getDepartmentById(Long id) throws Exception {
+        Department department = departmentDao.getDepartmentById(id);
+        if(department==null) {
+            throw new Exception("no data found");
+        }
+        return department;
+    }
+
+    public void updateDepartment(Department department) throws Exception {
+        Department departmentExist = departmentDao.getDepartmentById(department.getId());
+        if(departmentExist==null)
+            throw new Exception("no record found");
+
+        List<UserData> userDataList = userService.getAllUserByDepartment(department.getName());
+        departmentDao.save(department);
+        for(UserData userData: userDataList)
+        {
+            userService.updateUserByDepartment(userData.getId(),department.getName());
+        }
+
     }
 }
